feat: impl club registration button
